requirements.txt
MetaTrader5==5.0.4424
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.1
requests==2.31.0